Change highlight fov angle

+++ game_server/ServerParams.cpp
@@ -19,7 +19,7 @@ ServerParams::ServerParams() {
     playerCatRadius = config.GetFloat("PhysicsParams", "PlayerCatRadius", 0.5f);
     zombieRabbitMoveSpeed = config.GetFloat("PhysicsParams", "ZombieRabbitMoveSpeed", 1.6f);
     zombieRabbitRadius = config.GetFloat("PhysicsParams", "ZombieRabbitRadius", 0.5f);
-    highlightFOVAngle = PI / config.GetInteger("PhysicsParams", "HighlightFOVAngle", 3);
+    highlightFOVAngle = PI / config.GetFloat("PhysicsParams", "HighlightFOVAngle", 3);
     playerKnockBackMaxDistance = config.GetFloat("PhysicsParams", "PlayerKnockBackMaxDistance", 5.0f);
     playerRespawnInvincibleTime = config.GetFloat("PhysicsParams", "PlayerRespawnInvincibleTime", 2.0f);
